refactor(l5): move contract and contract-schemas from src/ into server/contracts (issue 864)
The contract builder and its schema table are server concerns; they sat in
src/ only because that is where they were first written. This moves the
implementation to server/contracts/ and leaves an L5 shim at each old path so
every existing importer keeps working until src/ retires at L6.

Moves, old -> new:

  src/contract.ts         -> server/contracts/contract.ts
  src/contract-schemas.ts -> server/contracts/contract-schemas.ts

contract.ts and contract-schemas.ts were each over the 500-code-line file
rule once relocated, so they are split by subject on the way in. New files,
all carved out of those two:

  server/contracts/contract-types.ts             types shared by the builder
  server/contracts/contract-capabilities.ts      capability assembly
  server/contracts/contract-payload-sections.ts  payload section assembly
  server/contracts/tool-contract.ts              the per-tool contract shape
  server/contracts/schemas-session.ts            session tool schemas
  server/contracts/schemas-events.ts             event tool schemas
  server/contracts/schemas-recall.ts             recall tool schemas
  server/contracts/schemas-repo-facts.ts         repo-fact tool schemas
  server/contracts/schemas-agent-surface.ts      agent-surface tool schemas

Shims left behind, each a single re-export and no logic:

  src/contract.ts
  src/contract-schemas.ts

Importers repointed off the old paths: server/contracts/declaration.ts,
server/contracts/declaration.test.ts, server/tools/get-contract.ts.

The split is a pure relocation. The emitted contract is byte-identical:
buildContract("2026-06-18T00:00:00.000Z").schema_hash is
4580a6813ae72bb67e4b45aa28dbfa4c3e49d685bcbd3fbd228d87ce3e3f4f7c both at
origin/main and on this branch, which is the value pinned at
src/contract.test.ts:21 and asserted again in declaration.test.ts.

src/contract.test.ts is deliberately left in src/, untouched and
byte-identical to origin/main, per rule M1: transforming it does not fit this
lane (it raises 61 TS2532 under noUncheckedIndexedAccess). It imports
./contract.ts, which is now the shim, so it still exercises the moved
implementation through it and is not staged here.

Verified:
  bunx tsc --noEmit -> exit 0
  oxlint --deny-warnings on all 16 staged paths -> exit 0
  bun run test:isolated src/contract.test.ts server/contracts/declaration.test.ts
    server/main.pg.test.ts -> 29 pass, 0 fail
  bash scripts/done-means/864-moved-out-of-src.sh -> judged=19 shims, PASS
  src-runtime-closure 50 -> 48

Refs #864
