+++ server/contracts/contract-payload-sections.ts
@@ -1,0 +1,292 @@
+/**
+ * The constant sections of the contract payload: interchange profiles, the
+ * agent memory adapter, the agent context pack, the receipt contract, and the
+ * promotion lifecycle.
+ *
+ * Split out of `buildContract` (issue 864) so that function stays within the
+ * size rule. These are literals with no inputs, spread back in unchanged, so
+ * the hashed payload and therefore the schema hash are byte-identical.
+ */
+export const CONTRACT_PAYLOAD_SECTIONS = {
+  interchange_profiles: {
+    okf: {
+      status: "compatibility-hooks" as const,
+      version: "draft" as const,
+      role: "edge-export-import-profile" as const,
+      metadata_path: "metadata.okf" as const,
+      reserved_files: ["index.md", "log.md"] as const,
+      required_frontmatter: ["type"] as const,
+      recommended_frontmatter: [
+        "title",
+        "description",
+        "resource",
+        "tags",
+        "timestamp",
+      ] as const,
+      export_surfaces: ["concept", "index", "log", "citations", "receipts"] as const,
+    },
+  },
+  agent_memory_adapter: {
+    status: "draft-local-contract" as const,
+    contract_doc: "docs/agent-memory-adapter-contract.md" as const,
+    server_authority: [
+      "auth",
+      "namespace",
+      "storage",
+      "promotion_policy",
+      "contract_discovery",
+    ] as const,
+    client_authority: [
+      "distillation",
+      "local_context",
+      "retry_spool",
+      "receipt_assembly",
+      "disclosure_export",
+    ] as const,
+    methods: {
+      start: {
+        maps_to: ["session_start", "lane_upsert"] as const,
+        owner: "client_and_server" as const,
+        status: "available" as const,
+      },
+      recall: {
+        maps_to: ["session_context", "search_all", "brain_answer"] as const,
+        owner: "client_and_server" as const,
+        status: "available" as const,
+      },
+      append_event: {
+        maps_to: ["append_session_event"] as const,
+        owner: "client_and_server" as const,
+        status: "available" as const,
+      },
+      compact: {
+        maps_to: ["session_context", "session_wrap"] as const,
+        owner: "client" as const,
+        status: "client-wrapper" as const,
+      },
+      wrap: {
+        maps_to: ["session_wrap"] as const,
+        owner: "client_and_server" as const,
+        status: "available" as const,
+      },
+      record_receipt: {
+        maps_to: ["append_session_event:event_type=receipt"] as const,
+        owner: "client_and_server" as const,
+        status: "client-wrapper" as const,
+      },
+      nominate_shared: {
+        maps_to: [
+          "append_session_event:metadata.share_candidate",
+          "append_session_event:metadata.memory_lifecycle_action=nominate_shared",
+        ] as const,
+        owner: "client_and_server" as const,
+        status: "available" as const,
+      },
+      candidate_memory: {
+        maps_to: [
+          "append_session_event:metadata.memory_lifecycle_action=candidate",
+        ] as const,
+        owner: "client" as const,
+        status: "client-wrapper" as const,
+      },
+      promote_candidate: {
+        maps_to: [
+          "append_session_event:metadata.memory_lifecycle_action=promote",
+        ] as const,
+        owner: "client" as const,
+        status: "client-wrapper" as const,
+      },
+      relegate_candidate: {
+        maps_to: [
+          "append_session_event:metadata.memory_lifecycle_action=relegate",
+        ] as const,
+        owner: "client" as const,
+        status: "client-wrapper" as const,
+      },
+      discard_candidate: {
+        maps_to: [
+          "append_session_event:metadata.memory_lifecycle_action=discard",
+        ] as const,
+        owner: "client" as const,
+        status: "client-wrapper" as const,
+      },
+      export_disclosure_bundle: {
+        maps_to: ["interchange_profiles.okf"] as const,
+        owner: "client" as const,
+        status: "client-wrapper" as const,
+      },
+    },
+  },
+  agent_context_pack: {
+    status: "runtime-available" as const,
+    availability: "mcp_tool_available" as const,
+    contract_doc: "docs/agent-context-pack-contract.md" as const,
+    parent_issue: 220 as const,
+    exact_scope_required: true as const,
+    scope_keys: [
+      "namespace",
+      "agent",
+      "platform",
+      "server_id",
+      "channel_id",
+      "thread_id",
+      "session_key",
+    ] as const,
+    sections: [
+      "working_set",
+      "recovery",
+      "durable_lane_context",
+      "durable_memory",
+      "profile_guidance",
+      "process_guidance",
+      "repo_facts",
+      "pointers",
+      "candidate_memory",
+    ] as const,
+    envelope_fields: ["warnings", "budget", "citations"] as const,
+    warning_fields: [
+      "missing_facts",
+      "stale_sources",
+      "degraded_sources",
+      "scope_denials",
+      "truncation",
+      "uncertainty",
+    ] as const,
+    working_set: {
+      status: "local-runtime-boundary" as const,
+      parent_issue: 222 as const,
+      implementation: "src/realtime/working-set.ts" as const,
+      storage: "ram_first_in_process" as const,
+      availability: "mcp_tool_available" as const,
+      item_label: "working_context" as const,
+      not_durable_memory: true as const,
+      exact_scope_required: true as const,
+      budget_defaults: {
+        ttl_ms: 1800000 as const,
+        max_sessions: 128 as const,
+        max_items_per_session: 24 as const,
+        max_global_items: 1024 as const,
+        max_item_chars: 4000 as const,
+        max_metadata_chars: 2000 as const,
+      },
+      counters: ["dropped", "expired", "trimmed"] as const,
+    },
+    recovery: {
+      status: "local-quarantine-boundary" as const,
+      parent_issue: 221 as const,
+      implementation: "src/realtime/recovery-wal.ts" as const,
+      storage: "env_configured_file_wal_with_in_memory_fallback" as const,
+      availability: "mcp_tool_available" as const,
+      item_label: "quarantined_recovery" as const,
+      not_durable_memory: true as const,
+      not_searchable_recall: true as const,
+      exact_scope_required: true as const,
+      explicit_include_required: true as const,
+      statuses: [
+        "active",
+        "wrapped",
+        "recovery_pending",
+        "reviewed",
+        "compacted",
+        "discarded",
+        "expired",
+      ] as const,
+      actions: [
+        "review",
+        "use_for_current_session",
+        "compact_to_wrap",
+        "promote_candidates",
+        "discard",
+        "defer",
+      ] as const,
+      budget_defaults: {
+        ttl_ms: 86400000 as const,
+        max_sessions: 128 as const,
+        max_items_per_session: 50 as const,
+        max_global_items: 2048 as const,
+        max_content_chars: 8000 as const,
+        max_metadata_chars: 2000 as const,
+        max_preview_chars: 1000 as const,
+      },
+      counters: ["dropped", "expired", "trimmed", "marked", "purged"] as const,
+    },
+    durable_lane_context: {
+      status: "runtime-available" as const,
+      implementation: "src/tools/agent-context-pack.ts" as const,
+      storage: "ob_session_lanes_and_events" as const,
+      availability: "mcp_tool_available" as const,
+      item_label: "durable_memory" as const,
+      exact_scope_required: true as const,
+      explicit_include_required: true as const,
+      scope_mismatch_behavior: "generic_scope_denial" as const,
+      budget_defaults: {
+        max_content_chars: 12000 as const,
+        max_context_chars: 6000 as const,
+        max_events: 8 as const,
+        max_event_chars: 1000 as const,
+      },
+    },
+  },
+  receipt_contract: {
+    status: "lightweight-openbrain-receipts" as const,
+    event_type: "receipt" as const,
+    contract_doc: "docs/agent-memory-adapter-contract.md" as const,
+    required_fields: [
+      "schema",
+      "action",
+      "agent",
+      "session_key",
+      "timestamp",
+      "sources",
+      "outputs",
+      "validations",
+    ] as const,
+    recommended_fields: [
+      "namespace",
+      "project",
+      "commands",
+      "external_channels",
+      "artifact_hashes",
+      "source_refs",
+      "residual_risk",
+    ] as const,
+    closed_brain_strict_fields: [
+      "preimage_hashes",
+      "postimage_hashes",
+      "base_document_hashes",
+      "tool_call_ids",
+      "approval_chain",
+      "redaction_policy",
+    ] as const,
+    secret_safe: true as const,
+  },
+  promotion_lifecycle: {
+    status: "explicit-client-owned-lifecycle" as const,
+    parent_issue: 224 as const,
+    contract_doc: "docs/agent-memory-adapter-contract.md" as const,
+    candidate_types: [
+      "user_preference",
+      "process_rule",
+      "channel_server_rule",
+      "code_repo_fact",
+      "positive_example",
+      "negative_example",
+      "durable_decision",
+      "shared_kb_nomination",
+    ] as const,
+    actions: [
+      "candidate",
+      "promote",
+      "relegate",
+      "discard",
+      "nominate_shared",
+    ] as const,
+    candidate_presence_effect: "no_durable_write_no_shared_write" as const,
+    shared_nomination_requires: [
+      "explicit_client_action",
+      "share_candidate_true",
+      "memory_lifecycle_action_nominate_shared",
+      "server_auth_scope_safety_provenance_secret_checks",
+    ] as const,
+  },
+} as const;
