+++ src/contract.ts
@@ -1,957 +1,2 @@
-import { createHash } from "node:crypto";
-import { TOOL_CONTRACTS } from "./contract-schemas.ts";
-
-// v24 (#678). The v23 string outlived three client-facing shape changes: the
-// live Zod schema for `agent_context_pack` gained `repo`, `prior_context`, and
-// `continue_from` while `src/contract-schemas.ts` — the hand-maintained mirror
-// the schema_hash is computed over — was never touched. So the version AND the
-// hash both read identically on either side of a real contract change, and
-// docs/downstream-rollout.md's "authoritative drift receipt" recorded nothing.
-// Bumping is the client-visible half; the hash moves on its own once the mirror
-// is honest, because TOOL_CONTRACTS is inside the hashed payload.
-export const CONTRACT_VERSION = "2026-08-09.memory-tools.v24";
-export const CONTRACT_SCHEMA_VERSION = 1;
-
-export interface ContractCapability {
-  name: string;
-  version: number;
-  kind: "tool" | "transport" | "schema";
-  description: string;
-}
-
-export interface OpenBrainContract {
-  service: "open-brain";
-  contract_version: string;
-  contract_scope: "required_openbrain_memory_contract";
-  schema_version: number;
-  schema_hash: string;
-  generated_at: string;
-  min_client_versions: Record<string, string>;
-  compatible_client_ranges: Record<string, string>;
-  transport: {
-    mcp: "streamable-http";
-    auth: "bearer";
-    namespace_boundary: "authorization";
-    session_required: true;
-  };
-  realtime_transport: {
-    nats_jetstream: {
-      status: "planned-transport-foundation" | "runtime-available";
-      availability: "available" | "not_runtime_available";
-      parent_issue: 223;
-      contract_doc: "docs/nats-jetstream-foundation.md";
-      server: {
-        planned_host: "production-host";
-        client_listen: "127.0.0.1:4222";
-        monitoring_listen: "127.0.0.1:8222";
-        jetstream_store_dir: "/opt/open-brain/nats/jetstream";
-      };
-      // Subjects are env-prefixed via the fleet-bus builder convention
-      // (src/nats-subjects.ts obContextPackSubject). The `{env}.` template below
-      // is a documentation placeholder; the live subject substitutes the slugged
-      // OPENBRAIN_NATS_ENV value (default "dev"), e.g. dev.ob.memory.context_pack.
-      subject_convention: "env_prefixed_fleet_bus";
-      request_reply_subjects: {
-        available: readonly ["{env}.ob.memory.context_pack"] | readonly [];
-        planned: readonly [
-          "{env}.ob.memory.session_start",
-          "{env}.ob.memory.append_event",
-          "{env}.ob.memory.wrap",
-          "{env}.ob.memory.resolve",
-          "{env}.ob.health",
-        ];
-      };
-      jetstream_streams: readonly [
-        "OB_AGENT_TRACE",
-        "OB_CONTEXT_PACK_REQUESTS",
-        "OB_CONTEXT_PACK_AUDIT",
-        "OB_PROMOTION_CANDIDATES",
-      ];
-      fallback_transport: "http_mcp";
-      auth_boundary: "openbrain_server_authority";
-      runtime_default: "http_mcp";
-    };
-  };
-  interchange_profiles: {
-    okf: {
-      status: "compatibility-hooks";
-      version: "draft";
-      role: "edge-export-import-profile";
-      metadata_path: "metadata.okf";
-      reserved_files: readonly ["index.md", "log.md"];
-      required_frontmatter: readonly ["type"];
-      recommended_frontmatter: readonly [
-        "title",
-        "description",
-        "resource",
-        "tags",
-        "timestamp",
-      ];
-      export_surfaces: readonly [
-        "concept",
-        "index",
-        "log",
-        "citations",
-        "receipts",
-      ];
-    };
-  };
-  agent_memory_adapter: {
-    status: "draft-local-contract";
-    contract_doc: "docs/agent-memory-adapter-contract.md";
-    server_authority: readonly [
-      "auth",
-      "namespace",
-      "storage",
-      "promotion_policy",
-      "contract_discovery",
-    ];
-    client_authority: readonly [
-      "distillation",
-      "local_context",
-      "retry_spool",
-      "receipt_assembly",
-      "disclosure_export",
-    ];
-    methods: Record<
-      string,
-      {
-        maps_to: readonly string[];
-        owner: "server" | "client" | "client_and_server";
-        status: "available" | "client-wrapper" | "planned";
-      }
-    >;
-  };
-  agent_context_pack: {
-    status: "runtime-available";
-    availability: "mcp_tool_available";
-    contract_doc: "docs/agent-context-pack-contract.md";
-    parent_issue: 220;
-    exact_scope_required: true;
-    scope_keys: readonly [
-      "namespace",
-      "agent",
-      "platform",
-      "server_id",
-      "channel_id",
-      "thread_id",
-      "session_key",
-    ];
-    sections: readonly [
-      "working_set",
-      "recovery",
-      "durable_lane_context",
-      "durable_memory",
-      "profile_guidance",
-      "process_guidance",
-      "repo_facts",
-      "pointers",
-      "candidate_memory",
-    ];
-    envelope_fields: readonly ["warnings", "budget", "citations"];
-    warning_fields: readonly [
-      "missing_facts",
-      "stale_sources",
-      "degraded_sources",
-      "scope_denials",
-      "truncation",
-      "uncertainty",
-    ];
-    working_set: {
-      status: "local-runtime-boundary";
-      parent_issue: 222;
-      implementation: "src/realtime/working-set.ts";
-      storage: "ram_first_in_process";
-      availability: "mcp_tool_available";
-      item_label: "working_context";
-      not_durable_memory: true;
-      exact_scope_required: true;
-      budget_defaults: {
-        ttl_ms: 1800000;
-        max_sessions: 128;
-        max_items_per_session: 24;
-        max_global_items: 1024;
-        max_item_chars: 4000;
-        max_metadata_chars: 2000;
-      };
-      counters: readonly ["dropped", "expired", "trimmed"];
-    };
-    recovery: {
-      status: "local-quarantine-boundary";
-      parent_issue: 221;
-      implementation: "src/realtime/recovery-wal.ts";
-      storage: "env_configured_file_wal_with_in_memory_fallback";
-      availability: "mcp_tool_available";
-      item_label: "quarantined_recovery";
-      not_durable_memory: true;
-      not_searchable_recall: true;
-      exact_scope_required: true;
-      explicit_include_required: true;
-      statuses: readonly [
-        "active",
-        "wrapped",
-        "recovery_pending",
-        "reviewed",
-        "compacted",
-        "discarded",
-        "expired",
-      ];
-      actions: readonly [
-        "review",
-        "use_for_current_session",
-        "compact_to_wrap",
-        "promote_candidates",
-        "discard",
-        "defer",
-      ];
-      budget_defaults: {
-        ttl_ms: 86400000;
-        max_sessions: 128;
-        max_items_per_session: 50;
-        max_global_items: 2048;
-        max_content_chars: 8000;
-        max_metadata_chars: 2000;
-        max_preview_chars: 1000;
-      };
-      counters: readonly ["dropped", "expired", "trimmed", "marked", "purged"];
-    };
-    durable_lane_context: {
-      status: "runtime-available";
-      implementation: "src/tools/agent-context-pack.ts";
-      storage: "ob_session_lanes_and_events";
-      availability: "mcp_tool_available";
-      item_label: "durable_memory";
-      exact_scope_required: true;
-      explicit_include_required: true;
-      scope_mismatch_behavior: "generic_scope_denial";
-      budget_defaults: {
-        max_content_chars: 12000;
-        max_context_chars: 6000;
-        max_events: 8;
-        max_event_chars: 1000;
-      };
-    };
-  };
-  receipt_contract: {
-    status: "lightweight-openbrain-receipts";
-    event_type: "receipt";
-    contract_doc: "docs/agent-memory-adapter-contract.md";
-    required_fields: readonly string[];
-    recommended_fields: readonly string[];
-    closed_brain_strict_fields: readonly string[];
-    secret_safe: true;
-  };
-  promotion_lifecycle: {
-    status: "explicit-client-owned-lifecycle";
-    parent_issue: 224;
-    contract_doc: "docs/agent-memory-adapter-contract.md";
-    candidate_types: readonly [
-      "user_preference",
-      "process_rule",
-      "channel_server_rule",
-      "code_repo_fact",
-      "positive_example",
-      "negative_example",
-      "durable_decision",
-      "shared_kb_nomination",
-    ];
-    actions: readonly [
-      "candidate",
-      "promote",
-      "relegate",
-      "discard",
-      "nominate_shared",
-    ];
-    candidate_presence_effect: "no_durable_write_no_shared_write";
-    shared_nomination_requires: readonly [
-      "explicit_client_action",
-      "share_candidate_true",
-      "memory_lifecycle_action_nominate_shared",
-      "server_auth_scope_safety_provenance_secret_checks",
-    ];
-  };
-  capabilities: ContractCapability[];
-  tool_contracts: Record<
-    string,
-    {
-      version: number;
-      input_schema: unknown;
-      output_shape: string;
-    }
-  >;
-}
-
-export const CONTRACT_CAPABILITIES: ContractCapability[] = [
-  {
-    name: "get_contract",
-    version: 1,
-    kind: "tool",
-    description: "Read the canonical Open Brain public contract manifest.",
-  },
-  {
-    name: "operator_doctor",
-    version: 1,
-    kind: "tool",
-    description:
-      "Read privileged operator doctor/status JSON for runtime, database, " +
-      "migrations, optional providers, transport, and log/audit health. " +
-      "Requires admin or ob-admin auth and never returns secrets or raw paths.",
-  },
-  {
-    name: "get_entry",
-    version: 2,
-    kind: "tool",
-    description:
-      "Fetch one readable memory row by table and UUID. Defaults to full row " +
-      "output; compact render returns a bounded exact-UUID preview envelope. " +
-      "Server-side auth and namespace predicates remain the security boundary " +
-      "for ID reads.",
-  },
-  {
-    name: "decompose_entry",
-    version: 1,
-    kind: "tool",
-    description:
-      "Plan dry-run-first decomposition of an oversized readable entry into " +
-      "smaller linked replacement thoughts. No writes occur unless the caller " +
-      "explicitly sets dry_run=false with apply_mode=write_replacements.",
-  },
-  {
-    name: "resolve_entry",
-    version: 1,
-    kind: "tool",
-    description:
-      "Resolve one UUID across readable Open Brain source families to its " +
-      "source type, namespace, and get_entry fetch path without semantic search.",
-  },
-  {
-    name: "upsert_repo_fact",
-    version: 2,
-    kind: "tool",
-    description:
-      "Record or update one curated, citation-backed fact about a code repository " +
-      "(qmd-derived) into graph entity metadata. Use to persist durable repo " +
-      "knowledge (how a module works, an invariant, a gotcha) with a source URL " +
-      "that proves it; re-upsert with the same key to correct an existing fact.",
-  },
-  {
-    name: "list_repo_facts",
-    version: 2,
-    kind: "tool",
-    description:
-      "Read back curated repository facts, scoped to your namespace and filtered " +
-      "by repo, collection, path, fact_type, or subject. Use to recall what is " +
-      "already known about a repo before re-deriving it.",
-  },
-  {
-    name: "log_thought",
-    version: 2,
-    kind: "tool",
-    description:
-      "Write a single durable thought, observation, or note to long-term memory. " +
-      "Use for free-form knowledge that is not tied to a session journal; it is " +
-      "embedded for later semantic search. For session-scoped events use " +
-      "append_session_event instead.",
-  },
-  {
-    name: "search_all",
-    version: 2,
-    kind: "tool",
-    description:
-      "Primary recall tool. Semantic + keyword search across Open Brain memory " +
-      "(thoughts, session events, facts) and optional qmd-backed code context. " +
-      "Call this before answering from assumption to ground yourself in stored " +
-      "knowledge.",
-  },
-  {
-    name: "session_start",
-    version: 2,
-    kind: "tool",
-    description:
-      "Open or resume a durable session lane and get recent events back. First-class " +
-      "callers supply agent/platform/server/channel/thread coordinates so the lane " +
-      "is established once and asserted scope mismatches fail closed.",
-  },
-  {
-    name: "session_context",
-    version: 3,
-    kind: "tool",
-    description:
-      "Read a session lane's current state and recent events without creating " +
-      "one. Use to rehydrate context for an existing session (by session_key or " +
-      "channel_id) before continuing work.",
-  },
-  {
-    name: "lane_upsert",
-    version: 2,
-    kind: "tool",
-    description:
-      "Create or update the metadata and rolling context of a session lane " +
-      "(status, project, agent, topic, current_context_md). Use to set or refresh " +
-      "the lane's high-level state; use append_session_event for individual " +
-      "journal entries.",
-  },
-  {
-    name: "lane_load",
-    version: 2,
-    kind: "tool",
-    description:
-      "List session lanes matching filters (key, project, agent, channel, " +
-      "status). Use to discover or pick up existing lanes; defaults to active " +
-      "lanes when status is omitted.",
-  },
-  {
-    name: "append_session_event",
-    version: 8,
-    kind: "tool",
-    description:
-      "Append one durable, typed event (fact, decision, blocker, action, etc.) " +
-      "to a session lane's journal. Supports first-write lane creation and " +
-      "atomically attaches supplied exact-scope coordinates to legacy lanes only " +
-      "where those coordinates were previously unasserted; conflicts fail closed.",
-  },
-  {
-    name: "citation_recall",
-    version: 1,
-    kind: "tool",
-    description:
-      "Read a session event's stored host-neutral transcript citation and " +
-      "bounded neighboring exchanges, or explicitly report source_not_stored " +
-      "for legacy evidence-less events.",
-  },
-  {
-    name: "session_wrap",
-    version: 2,
-    kind: "tool",
-    description:
-      "Checkpoint a session lane with a durable summary, key decisions, and next " +
-      "steps. Exact scope is validated before the session record and lane " +
-      "current_context_md are updated transactionally for immediate recall.",
-  },
-  {
-    name: "entity_graph",
-    version: 2,
-    kind: "schema",
-    description:
-      "Open Brain graph entities and links, including archived entity lifecycle.",
-  },
-  {
-    name: "session_lanes",
-    version: 2,
-    kind: "schema",
-    description:
-      "Durable session lanes, events, context, wraps, and host-neutral transcript citations.",
-  },
-  // v3 since #678, when the published tool contract gained `repo`,
-  // `prior_context`, and `continue_from`. `tool_contracts` moved
-  // (contract-schemas.ts) and this capabilities entry did not, so the
-  // manifest advertised 3 in one block and 2 in the other. The Python client
-  // validates BOTH against FIRST_CLASS_RUNTIME_TOOL_VERSIONS (client.py:78)
-  // and refuses a v2 manifest by design, so every direct capture returned
-  // `status: lost` -- silently, since the runtime receipt reports the failure
-  // and nothing was watching it. That is the same hand-maintained-mirror
-  // drift this file's header comment describes.
-  //
-  // Keep name/version/kind on adjacent lines: the Python client's
-  // test_required_contract_matches_server_source_of_truth parses this file
-  // with a regex that expects them contiguous.
-  {
-    name: "agent_context_pack",
-    version: 3,
-    kind: "tool",
-    description:
-      "First-class realtime context-pack tool for Hermes and future agents. " +
-      "It exposes exact-scope RAM working_set, explicitly opted-in quarantined " +
-      "recovery, and explicitly requested bounded durable_lane_context over MCP; " +
-      "the optional NATS bridge returns the same server-authoritative pack.",
-  },
-  {
-    name: "agent_reflex_pointers",
-    version: 1,
-    kind: "tool",
-    description:
-      "Per-turn reflex projection over the single agent_context_pack durable " +
-      "recall and pointer machinery. Returns budget-bounded, body-free, cited " +
-      "resolvable pointers to durable records relevant to the current query, with " +
-      "prior-context suppression applied; placement into the model prompt stays " +
-      "client-owned and the tool returns an ordinary result envelope only.",
-  },
-  {
-    name: "working_set_append",
-    version: 1,
-    kind: "tool",
-    description:
-      "Append one RAM-only working-context item for an exact active " +
-      "namespace/agent/platform/server/channel/thread/session scope. This " +
-      "does not create durable memory or shared-kb rows.",
-  },
-  {
-    name: "recovery_wal_append",
-    version: 1,
-    kind: "tool",
-    description:
-      "Append one exact-scope quarantined recovery WAL record for interrupted " +
-      "agent traces. Recovery WAL records are unreviewed, not durable memory, " +
-      "and not searchable recall.",
-  },
-  {
-    name: "recovery_wal_mark",
-    version: 1,
-    kind: "tool",
-    description:
-      "Mark or purge one exact-scope quarantined recovery WAL record after " +
-      "review without promoting it into durable memory.",
-  },
-  {
-    name: "agent_memory_adapter",
-    version: 1,
-    kind: "schema",
-    description:
-      "Draft local adapter contract for agent memory lifecycle clients. " +
-      "Defines start, recall, append_event, compact, wrap, record_receipt, " +
-      "nominate_shared, and export_disclosure_bundle without moving server " +
-      "auth or namespace authority into clients.",
-  },
-  {
-    name: "receipt_contract",
-    version: 1,
-    kind: "schema",
-    description:
-      "Lightweight citation-safe receipt metadata model for Open Brain " +
-      "session events, with stricter Closed Brain fields marked separately.",
-  },
-  {
-    name: "memory_promotion_lifecycle",
-    version: 1,
-    kind: "schema",
-    description:
-      "Explicit client-owned lifecycle for candidate memory, durable promotion, " +
-      "relegation/discard, and shared-kb nomination. Candidate presence alone " +
-      "does not write durable memory or shared-kb.",
-  },
-  {
-    name: "streamable_http_auth",
-    version: 1,
-    kind: "transport",
-    description:
-      "Bearer-token identity establishes namespace boundaries for MCP sessions.",
-  },
-];
-
-function sortValue(value: unknown): unknown {
-  if (Array.isArray(value)) return value.map(sortValue);
-  if (value && typeof value === "object") {
-    return Object.fromEntries(
-      Object.entries(value as Record<string, unknown>)
-        .sort(([a], [b]) => a.localeCompare(b))
-        .map(([key, val]) => [key, sortValue(val)]),
-    );
-  }
-  return value;
-}
-
-export function stableJson(value: unknown): string {
-  return JSON.stringify(sortValue(value));
-}
-
-function requiredContractHashPayload(
-  payload: Omit<OpenBrainContract, "generated_at" | "schema_hash">,
-): Omit<
-  OpenBrainContract,
-  "generated_at" | "schema_hash" | "realtime_transport"
-> {
-  const { realtime_transport: _advisoryRealtimeTransport, ...requiredPayload } =
-    payload;
-  return requiredPayload;
-}
-
-export function contractHash(
-  payload: Omit<OpenBrainContract, "generated_at" | "schema_hash">,
-): string {
-  return createHash("sha256")
-    .update(stableJson(requiredContractHashPayload(payload)))
-    .digest("hex");
-}
-
-export function buildContract(
-  generatedAt = new Date().toISOString(),
-  options: {
-    natsAvailability?: "available" | "not_runtime_available";
-  } = {},
-): OpenBrainContract {
-  const natsAvailability = options.natsAvailability ?? "not_runtime_available";
-  const payload = {
-    service: "open-brain" as const,
-    contract_version: CONTRACT_VERSION,
-    contract_scope: "required_openbrain_memory_contract" as const,
-    schema_version: CONTRACT_SCHEMA_VERSION,
-    min_client_versions: {
-      "openbrain-memory": "0.1.15",
-      "rtech-hermes-runtime": "0.1.0",
-      mcp2cli: "0.3.6",
-    },
-    compatible_client_ranges: {
-      "openbrain-memory": ">=0.1.15 <1.0.0",
-      "rtech-hermes-runtime": ">=0.1.0 <1.0.0",
-      mcp2cli: ">=0.3.6 <1.0.0",
-    },
-    transport: {
-      mcp: "streamable-http" as const,
-      auth: "bearer" as const,
-      namespace_boundary: "authorization" as const,
-      session_required: true as const,
-    },
-    // ADVISORY: realtime_transport is EXCLUDED from schema_hash
-    // (see requiredContractHashPayload — it destructures realtime_transport out
-    // before hashing). It is documentation of the NATS foundation, so its
-    // contents (subject shape, availability, streams) can change without a
-    // schema_hash bump or a contract-version break.
-    realtime_transport: {
-      nats_jetstream: {
-        status:
-          natsAvailability === "available"
-            ? ("runtime-available" as const)
-            : ("planned-transport-foundation" as const),
-        availability: natsAvailability,
-        parent_issue: 223 as const,
-        contract_doc: "docs/nats-jetstream-foundation.md" as const,
-        server: {
-          planned_host: "production-host" as const,
-          client_listen: "127.0.0.1:4222" as const,
-          monitoring_listen: "127.0.0.1:8222" as const,
-          jetstream_store_dir:
-            "/opt/open-brain/nats/jetstream" as const,
-        },
-        // Env-prefixed subjects (fleet-bus convention). The advertised strings
-        // carry a `{env}.` template placeholder; the runtime substitutes the
-        // slugged OPENBRAIN_NATS_ENV value via obContextPackSubject(env).
-        subject_convention: "env_prefixed_fleet_bus" as const,
-        request_reply_subjects: {
-          available:
-            natsAvailability === "available"
-              ? (["{env}.ob.memory.context_pack"] as const)
-              : ([] as const),
-          planned: [
-            "{env}.ob.memory.session_start",
-            "{env}.ob.memory.append_event",
-            "{env}.ob.memory.wrap",
-            "{env}.ob.memory.resolve",
-            "{env}.ob.health",
-          ] as const,
-        },
-        jetstream_streams: [
-          "OB_AGENT_TRACE",
-          "OB_CONTEXT_PACK_REQUESTS",
-          "OB_CONTEXT_PACK_AUDIT",
-          "OB_PROMOTION_CANDIDATES",
-        ] as const,
-        fallback_transport: "http_mcp" as const,
-        auth_boundary: "openbrain_server_authority" as const,
-        runtime_default: "http_mcp" as const,
-      },
-    },
-    interchange_profiles: {
-      okf: {
-        status: "compatibility-hooks" as const,
-        version: "draft" as const,
-        role: "edge-export-import-profile" as const,
-        metadata_path: "metadata.okf" as const,
-        reserved_files: ["index.md", "log.md"] as const,
-        required_frontmatter: ["type"] as const,
-        recommended_frontmatter: [
-          "title",
-          "description",
-          "resource",
-          "tags",
-          "timestamp",
-        ] as const,
-        export_surfaces: [
-          "concept",
-          "index",
-          "log",
-          "citations",
-          "receipts",
-        ] as const,
-      },
-    },
-    agent_memory_adapter: {
-      status: "draft-local-contract" as const,
-      contract_doc: "docs/agent-memory-adapter-contract.md" as const,
-      server_authority: [
-        "auth",
-        "namespace",
-        "storage",
-        "promotion_policy",
-        "contract_discovery",
-      ] as const,
-      client_authority: [
-        "distillation",
-        "local_context",
-        "retry_spool",
-        "receipt_assembly",
-        "disclosure_export",
-      ] as const,
-      methods: {
-        start: {
-          maps_to: ["session_start", "lane_upsert"] as const,
-          owner: "client_and_server" as const,
-          status: "available" as const,
-        },
-        recall: {
-          maps_to: ["session_context", "search_all", "brain_answer"] as const,
-          owner: "client_and_server" as const,
-          status: "available" as const,
-        },
-        append_event: {
-          maps_to: ["append_session_event"] as const,
-          owner: "client_and_server" as const,
-          status: "available" as const,
-        },
-        compact: {
-          maps_to: ["session_context", "session_wrap"] as const,
-          owner: "client" as const,
-          status: "client-wrapper" as const,
-        },
-        wrap: {
-          maps_to: ["session_wrap"] as const,
-          owner: "client_and_server" as const,
-          status: "available" as const,
-        },
-        record_receipt: {
-          maps_to: ["append_session_event:event_type=receipt"] as const,
-          owner: "client_and_server" as const,
-          status: "client-wrapper" as const,
-        },
-        nominate_shared: {
-          maps_to: [
-            "append_session_event:metadata.share_candidate",
-            "append_session_event:metadata.memory_lifecycle_action=nominate_shared",
-          ] as const,
-          owner: "client_and_server" as const,
-          status: "available" as const,
-        },
-        candidate_memory: {
-          maps_to: [
-            "append_session_event:metadata.memory_lifecycle_action=candidate",
-          ] as const,
-          owner: "client" as const,
-          status: "client-wrapper" as const,
-        },
-        promote_candidate: {
-          maps_to: [
-            "append_session_event:metadata.memory_lifecycle_action=promote",
-          ] as const,
-          owner: "client" as const,
-          status: "client-wrapper" as const,
-        },
-        relegate_candidate: {
-          maps_to: [
-            "append_session_event:metadata.memory_lifecycle_action=relegate",
-          ] as const,
-          owner: "client" as const,
-          status: "client-wrapper" as const,
-        },
-        discard_candidate: {
-          maps_to: [
-            "append_session_event:metadata.memory_lifecycle_action=discard",
-          ] as const,
-          owner: "client" as const,
-          status: "client-wrapper" as const,
-        },
-        export_disclosure_bundle: {
-          maps_to: ["interchange_profiles.okf"] as const,
-          owner: "client" as const,
-          status: "client-wrapper" as const,
-        },
-      },
-    },
-    agent_context_pack: {
-      status: "runtime-available" as const,
-      availability: "mcp_tool_available" as const,
-      contract_doc: "docs/agent-context-pack-contract.md" as const,
-      parent_issue: 220 as const,
-      exact_scope_required: true as const,
-      scope_keys: [
-        "namespace",
-        "agent",
-        "platform",
-        "server_id",
-        "channel_id",
-        "thread_id",
-        "session_key",
-      ] as const,
-      sections: [
-        "working_set",
-        "recovery",
-        "durable_lane_context",
-        "durable_memory",
-        "profile_guidance",
-        "process_guidance",
-        "repo_facts",
-        "pointers",
-        "candidate_memory",
-      ] as const,
-      envelope_fields: ["warnings", "budget", "citations"] as const,
-      warning_fields: [
-        "missing_facts",
-        "stale_sources",
-        "degraded_sources",
-        "scope_denials",
-        "truncation",
-        "uncertainty",
-      ] as const,
-      working_set: {
-        status: "local-runtime-boundary" as const,
-        parent_issue: 222 as const,
-        implementation: "src/realtime/working-set.ts" as const,
-        storage: "ram_first_in_process" as const,
-        availability: "mcp_tool_available" as const,
-        item_label: "working_context" as const,
-        not_durable_memory: true as const,
-        exact_scope_required: true as const,
-        budget_defaults: {
-          ttl_ms: 1800000 as const,
-          max_sessions: 128 as const,
-          max_items_per_session: 24 as const,
-          max_global_items: 1024 as const,
-          max_item_chars: 4000 as const,
-          max_metadata_chars: 2000 as const,
-        },
-        counters: ["dropped", "expired", "trimmed"] as const,
-      },
-      recovery: {
-        status: "local-quarantine-boundary" as const,
-        parent_issue: 221 as const,
-        implementation: "src/realtime/recovery-wal.ts" as const,
-        storage: "env_configured_file_wal_with_in_memory_fallback" as const,
-        availability: "mcp_tool_available" as const,
-        item_label: "quarantined_recovery" as const,
-        not_durable_memory: true as const,
-        not_searchable_recall: true as const,
-        exact_scope_required: true as const,
-        explicit_include_required: true as const,
-        statuses: [
-          "active",
-          "wrapped",
-          "recovery_pending",
-          "reviewed",
-          "compacted",
-          "discarded",
-          "expired",
-        ] as const,
-        actions: [
-          "review",
-          "use_for_current_session",
-          "compact_to_wrap",
-          "promote_candidates",
-          "discard",
-          "defer",
-        ] as const,
-        budget_defaults: {
-          ttl_ms: 86400000 as const,
-          max_sessions: 128 as const,
-          max_items_per_session: 50 as const,
-          max_global_items: 2048 as const,
-          max_content_chars: 8000 as const,
-          max_metadata_chars: 2000 as const,
-          max_preview_chars: 1000 as const,
-        },
-        counters: [
-          "dropped",
-          "expired",
-          "trimmed",
-          "marked",
-          "purged",
-        ] as const,
-      },
-      durable_lane_context: {
-        status: "runtime-available" as const,
-        implementation: "src/tools/agent-context-pack.ts" as const,
-        storage: "ob_session_lanes_and_events" as const,
-        availability: "mcp_tool_available" as const,
-        item_label: "durable_memory" as const,
-        exact_scope_required: true as const,
-        explicit_include_required: true as const,
-        scope_mismatch_behavior: "generic_scope_denial" as const,
-        budget_defaults: {
-          max_content_chars: 12000 as const,
-          max_context_chars: 6000 as const,
-          max_events: 8 as const,
-          max_event_chars: 1000 as const,
-        },
-      },
-    },
-    receipt_contract: {
-      status: "lightweight-openbrain-receipts" as const,
-      event_type: "receipt" as const,
-      contract_doc: "docs/agent-memory-adapter-contract.md" as const,
-      required_fields: [
-        "schema",
-        "action",
-        "agent",
-        "session_key",
-        "timestamp",
-        "sources",
-        "outputs",
-        "validations",
-      ] as const,
-      recommended_fields: [
-        "namespace",
-        "project",
-        "commands",
-        "external_channels",
-        "artifact_hashes",
-        "source_refs",
-        "residual_risk",
-      ] as const,
-      closed_brain_strict_fields: [
-        "preimage_hashes",
-        "postimage_hashes",
-        "base_document_hashes",
-        "tool_call_ids",
-        "approval_chain",
-        "redaction_policy",
-      ] as const,
-      secret_safe: true as const,
-    },
-    promotion_lifecycle: {
-      status: "explicit-client-owned-lifecycle" as const,
-      parent_issue: 224 as const,
-      contract_doc: "docs/agent-memory-adapter-contract.md" as const,
-      candidate_types: [
-        "user_preference",
-        "process_rule",
-        "channel_server_rule",
-        "code_repo_fact",
-        "positive_example",
-        "negative_example",
-        "durable_decision",
-        "shared_kb_nomination",
-      ] as const,
-      actions: [
-        "candidate",
-        "promote",
-        "relegate",
-        "discard",
-        "nominate_shared",
-      ] as const,
-      candidate_presence_effect: "no_durable_write_no_shared_write" as const,
-      shared_nomination_requires: [
-        "explicit_client_action",
-        "share_candidate_true",
-        "memory_lifecycle_action_nominate_shared",
-        "server_auth_scope_safety_provenance_secret_checks",
-      ] as const,
-    },
-    capabilities: [...CONTRACT_CAPABILITIES].sort((a, b) =>
-      a.name.localeCompare(b.name),
-    ),
-    tool_contracts: TOOL_CONTRACTS,
-  };
-
-  return {
-    ...payload,
-    schema_hash: contractHash(payload),
-    generated_at: generatedAt,
-  };
-}
+// L5 shim (issue 864): moved to server/contracts/contract.ts; retired with src/ at L6.
+export * from "../server/contracts/contract.ts";
